common: Move global configuration state modifiers back to settings
