+++ src/common/settings_common.cpp
@@ -52,14 +52,4 @@ const std::string& BasicSetting::GetLabel() const {
     return label;
 }
 
-static bool configuring_global = true;
-
-bool IsConfiguringGlobal() {
-    return configuring_global;
-}
-
-void SetConfiguringGlobal(bool is_global) {
-    configuring_global = is_global;
-}
-
 } // namespace Settings
